rework(auth): rename token.service && implement /auth/logout api route
