+++ .env.examples
@@ -11,4 +11,5 @@ SQL_DB_DATABASE=
 
 MONGO_DB_URI=
 
-JWT_SECRET=
+JWT_ACCESS_TOKEN_SECRET=
+JWT_ACCESS_TOKEN_EXPIRATION_TIME=
