+++ src/modules/auth/auth.service.ts
@@ -1,14 +1,14 @@
-import { Injectable } from '@nestjs/common';
+import { HttpStatus, Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 
 import { UserService } from '../user/user.service';
-import { RefreshTokenService } from './token.service';
+import { RefreshTokenService } from './refresh-token.service';
 
 @Injectable()
 export class AuthService {
   constructor(
     private userService: UserService,
-    private tokenService: RefreshTokenService,
+    private refreshTokenService: RefreshTokenService,
     private jwtService: JwtService,
   ) {}
 
@@ -26,8 +26,9 @@ export class AuthService {
 
   async login(user: any) {
     const payload = { username: user.name, uid: user.id };
+
     const token = this.jwtService.sign(payload);
-    const refreshToken = await this.tokenService.issueRefreshToken(user.id);
+    const refreshToken = await this.refreshTokenService.create(user.id, 30);
 
     return {
       token,
@@ -40,7 +41,19 @@ export class AuthService {
     // TODO: CREATE USER WITH USER.SERVICE
   }
 
+  async logout(userId) {
+    const { affected: isTokensDeleted } = await this.refreshTokenService.delete(
+      userId,
+    );
+
+    if (isTokensDeleted) {
+      return HttpStatus.OK;
+    }
+
+    return HttpStatus.NO_CONTENT;
+  }
+
   async refresh(token: string) {
-    return this.tokenService.verifyRefreshToken(token);
+    return this.refreshTokenService.verify(token);
   }
 }
